Check if file exists before deleting it in segment_all.rb

+++ ruby/segment_all.rb
@@ -100,7 +100,10 @@ SKIP_WINDOW_OPTIONS.each do |skip_window_option|
                    "window-#{window_size}_filter-#{filter}.log"
 
             next if skip_existing && File.file?(file)
-            File.delete(file)
+            
+			if File.file?(file)
+			  File.delete(file)
+			end
 
             CSV_FILES.each do |csv_file|
               run_string = "#{script} "\
